backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-archive-backend"
version = "0.1.0"
description = "Document archive backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.1.19",
  "alembic>=1.13.2",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.4",
  "python-multipart>=0.0.9",
  "celery>=5.4.0",
  "redis>=5.0.7",
  "minio>=7.2.7",
  "httpx>=0.27.0",
  "python-dateutil>=2.9.0",
  "structlog>=24.2.0",
  "prometheus-client>=0.20.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.0,<4.1",
  "itsdangerous>=2.2.0",
  "pypdf>=4.3.0",
  "python-docx>=1.1.2"
]

[project.optional-dependencies]
test = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "testcontainers[postgres,redis]>=4.7.1"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]
